refactor: Move lint specific tests to lints/mod.rs
